RPET-774 refactor validation and update tests
