+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/functionaltest/MigrateChequePaymentTest.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.divorce.orchestration.functionaltest;
 
-import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.matching.EqualToPattern;
 import org.junit.Before;
 import org.junit.Test;
@@ -23,7 +22,6 @@ import java.util.Map;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.jayway.jsonpath.matchers.JsonPathMatchers.hasJsonPath;
 import static com.jayway.jsonpath.matchers.JsonPathMatchers.isJson;
 import static org.hamcrest.Matchers.allOf;
@@ -31,7 +29,6 @@ import static org.hamcrest.Matchers.emptyOrNullString;
 import static org.hamcrest.core.Is.is;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
-import static org.springframework.http.HttpMethod.POST;
 import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -86,13 +83,6 @@ public class MigrateChequePaymentTest extends IdamTestSupport {
             .eventId(EVENT_ID)
             .caseDetails(caseDetails)
             .build();
-
-        maintenanceServiceServer.stubFor(WireMock.request(POST.name(),urlEqualTo(
-            "/casemaintenance/version/1/updateCase/test.case.id/MigrateChequePayment"
-        )).willReturn(aResponse()
-                .withStatus(HttpStatus.OK.value())
-                .withHeader(CONTENT_TYPE, APPLICATION_JSON_VALUE)
-                .withBody(convertObjectToJsonString(migratedCaseData))));
 
         webClient.perform(post(API_URL)
             .content(convertObjectToJsonString(request))
